refactor: reorganize project into src app structure
