+++ __tests__/home.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen } from '@testing-library/react';
-import Home from '../pages/index';
+import Home from '../src/app/page';
 
 jest.mock('swr', () => () => ({ data: [{ id: 1, name: 'Product', price: 1, color: 'Red' }] }));
