+++ src/app/products/[id]/page.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { useRouter } from 'next/router';
 import useSWR from 'swr';
-import Layout from '../../components/Layout';
-import { useCart } from '../../lib/cartContext';
+import Layout from '../../../components/Layout';
+import { useCart } from '../../../lib/cartContext';
 /**
  * Product detail page that displays information about a single product
  * and allows the user to add it to their cart.
